Listo re listo negocio. Empezando persistencia

+++ src/negocio/VORestaurante.java
@@ -19,7 +19,7 @@ public interface VORestaurante {
 	 * 
 	 * @return la capacidad
 	 */
-	public Long getCapacidad();
+	public Integer getCapacidad();
 	
 	/**
 	 *
